# internal/proto/main.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "main.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Booking"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/page/get/{list}": {
      "get": {
        "operationId": "Booking_GetList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/helloRespListPage"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "list",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Booking"
        ]
      }
    },
    "/api/product/get/{name}": {
      "get": {
        "operationId": "Booking_GetProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/helloproduct"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Booking"
        ]
      }
    }
  },
  "definitions": {
    "helloRespListPage": {
      "type": "object",
      "properties": {
        "elements": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/helloelement"
          }
        }
      }
    },
    "helloelement": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "price": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "helloproduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}